references/references_requirements.txt
biopython
jsonschema
kallisto
snakemake
bowtie2
star
hisat2
